Make game work with pygbag
